chore(metadata): added metadata on static/dynamic pages
